+++ src/lib/ngx-stars.component.ts
@@ -1,11 +1,11 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 
 @Component({
   selector: 'ngx-stars',
   templateUrl: './ngx-stars.component.html',
   styleUrls: [ './ngx-stars.component.css' ],
 })
-export class NgxStarsComponent implements OnInit {
+export class NgxStarsComponent implements OnInit, OnDestroy {
 
   @Input()
   maxStars: number = 5;
@@ -34,6 +34,9 @@ export class NgxStarsComponent implements OnInit {
   @Input()
   wholeStars: boolean = false;
 
+  @Input()
+  customStarIcons: { empty: string, half: string, full: string };
+
   @Output()
   ratingOutput: EventEmitter<number> = new EventEmitter();
 
@@ -42,7 +45,11 @@ export class NgxStarsComponent implements OnInit {
   animationInterval: any;
   animationRunning: boolean;
 
+  private customCssClasses: HTMLStyleElement[];
+  private customClassIdentifier = Math.random().toString(36).substring(2);
+
   ngOnInit(): void {
+    this.setupStarImages();
     this.editableStars = Array.from(new Array(this.maxStars)).map((elem, index) => new EditableStar(index));
     this.setRating(this.initialStars);
 
@@ -51,29 +58,58 @@ export class NgxStarsComponent implements OnInit {
     }
   }
 
-  starColorAndPadding(): Object {
-    return Object.assign({}, this.starColor(), this.starPadding());
-  }
-
-  private starColor(): Object {
-    return { color: this.color || 'crimson' };
-  }
-
-  private starPadding(): Object {
-    let padding = '0.5rem';
-    if (Number.isInteger(this.size) || this.size > 0 || this.size < 6) {
-      padding = `0.${this.size}rem`;
+  ngOnDestroy(): void {
+    // remove the three custom classes we created if custom image urls were provided
+    if (this.customCssClasses) {
+      this.customCssClasses.forEach(style => {
+        if (style && style.parentNode) {
+          style.parentNode.removeChild(style);
+        }
+      });
     }
-
-    return { 'margin-right': this.customPadding || padding };
   }
 
-  starSize(): string {
-    if (!Number.isInteger(this.size) || this.size < 2 || this.size > 5) {
-      return '';
+  private setupStarImages() {
+    if (this.customStarIcons) {
+      this.customCssClasses = [];
+      Object.keys(this.customStarIcons).map(key => key as StarType).forEach(starType => {
+        const classname = this.getStarClass(starType);
+        this.createCssClass(classname, starType);
+      });
     }
-    return `fa-${this.size}x`;
   }
+
+  private createCssClass(classname: string, starType: StarType) {
+    const clazz = document.createElement('style');
+    clazz.type = 'text/css';
+    clazz.innerHTML = `.${classname} {
+      -webkit-mask-image: url(${this.customStarIcons[starType]});
+      mask-image: url(${this.customStarIcons[starType]});
+    }`;
+    document.getElementsByTagName('head')[0].appendChild(clazz);
+    this.customCssClasses.push(clazz);
+  }
+
+  starPadding(): { [p: string]: string } {
+    return { 'margin-right': this.customPadding || `0.${this.safeSize()}rem` };
+  }
+
+  starColorAndSize(): { [p: string]: string } {
+    return Object.assign({}, this.starColor(), this.starSize());
+  }
+
+  private starColor(): { [p: string]: string } {
+    return { 'background-color': this.color || 'crimson' };
+  }
+
+  starSize(): { [p: string]: string } {
+    return {
+      height: `${15 * this.safeSize()}px`,
+      width: `${16 * this.safeSize()}px`,
+    };
+  }
+
+  private safeSize = () => (Number.isInteger(this.size) && this.size > 0 && this.size < 6) ? this.size : 1;
 
   starAnimation(): void {
     this.animationRunning = true;
@@ -104,15 +140,15 @@ export class NgxStarsComponent implements OnInit {
     const clickedInFirstHalf = this.clickedInFirstHalf(event);
 
     // fill in either a half or whole star depending on where user clicked
-    clickedStar.classname = (!this.wholeStars && clickedInFirstHalf) ? 'fa-star-half-o' : 'fa-star';
+    clickedStar.classname = (!this.wholeStars && clickedInFirstHalf) ? this.getStarClass('half') : this.getStarClass('full');
 
     // fill in all stars in previous positions and clear all in later ones
     this.editableStars.forEach(star => {
       if (star.position > clickedStar.position) {
-        star.classname = 'fa-star-o';
+        star.classname = this.getStarClass('empty');
       }
       else if (star.position < clickedStar.position) {
-        star.classname = 'fa-star';
+        star.classname = this.getStarClass('full');
       }
     });
   }
@@ -134,7 +170,7 @@ export class NgxStarsComponent implements OnInit {
 
   onZeroStarHover(): void {
     // clear all stars
-    this.editableStars.forEach(star => star.classname = 'fa-star-o');
+    this.editableStars.forEach(star => star.classname = this.getStarClass('empty'));
   }
 
   onStarsUnhover() {
@@ -142,13 +178,13 @@ export class NgxStarsComponent implements OnInit {
     this.editableStars.forEach(star => {
       const starNumber = star.position + 1;
       if (this.rating >= starNumber) {
-        star.classname = 'fa-star';
+        star.classname = this.getStarClass('full');
       }
       else if (this.rating > starNumber - 1 && this.rating < starNumber) {
-        star.classname = 'fa-star-half-o';
+        star.classname = this.getStarClass('half');
       }
       else {
-        star.classname = 'fa-star-o';
+        star.classname = this.getStarClass('empty');
       }
     });
   }
@@ -159,7 +195,16 @@ export class NgxStarsComponent implements OnInit {
   }
 
   noop(): void {}
+
+  private getStarClass(starType: StarType) {
+    if (this.customCssClasses) {
+      return `ngx-stars-star-${starType}-${this.customClassIdentifier}`;
+    }
+    return `star-${starType}`;
+  }
 }
+
+export type StarType = 'empty' | 'half' | 'full';
 
 export class EditableStar {
   position: number;
@@ -167,6 +212,5 @@ export class EditableStar {
 
   constructor(position: number) {
     this.position = position;
-    this.classname = 'fa-star-o';
   }
 }